requirements.txt
asgiref==3.2.3
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
confluent-kafka==1.4.2
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-createsuperuser==0.0.0
django-filter==2.2.0
django-heroku==0.3.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
joblib==0.14.1
lxml==4.5.0
Markdown==3.1.1
numpy==1.18.4
pandas==1.0.3
parsel==1.5.2
Protego==0.1.16
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-scrapyd-api==2.1.2
pytz==2019.3
queuelib==1.5.0
requests==2.23.0
s2geometry==0.9.0
scikit-learn==0.22
scipy==1.4.1
Scrapy==2.0.1
scrapy-djangoitem==1.1.1
scrapyd==1.2.1
service-identity==18.1.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
sqlparse==0.3.0
SVC==0.1
threadpoolctl==2.0.0
Twisted==20.3.0
urllib3==1.25.8
virtualenv==16.7.9
w3lib==1.21.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.0.1
